удаляет моки
